Force conversion of raw data to floats. If that's not possible, an error is raised straight away. w/o this, you get much more cryptic error messages later

+++ superplot/data_loader.py
@@ -49,7 +49,7 @@ def _read_data_file(file_name):
     data_frame = data_frame.transpose()
 
     # Find array from data-frame
-    data_array = data_frame.values
+    data_array = data_frame.values.astype('float64')
 
     return data_array
